requirements.txt
tensorflow==2.15.0
opencv-python==4.8.0.76
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
pyyaml==6.0
